requirements.txt
llama-index-core==0.11.23
qdrant_client==1.10.1
langchain-core==0.3.18